refactor: named structs/helpers over tuples; split shows_chrome
- SenderStatusRow converted from a tuple type alias to a named struct; query_map and tests updated
- Split shows_chrome into shows_left_panel / shows_global_search; OnboardingRelay keeps search only
- Extract sync_unlock_window_state in main.rs; add an unlock-window gating test
- compose_window: HeaderRenderOutput + recipient-chip/label helpers extracted
- Drop BadgeKind enum; rename paint_badge_at to paint_nostr_badge_at
- Test hygiene: assert_eq!(x,true)->assert!(x); first().is_some()->!is_empty(); relocate refresh-plan tests

+++ crates/hoot-backend/src/db/queries.rs
@@ -246,7 +246,7 @@ ORDER BY le.created_at DESC
                )",
         )?;
 
-        let mail_kind = u32::from(MAIL_EVENT_KIND as u16);
+        let mail_kind = u32::from(MAIL_EVENT_KIND);
 
         let id_iter = stmt.query_map([mail_kind], |row| {
             let id: String = row.get(0)?;
